+++ aeron-client/src/main/c/concurrent/aeron_spsc_concurrent_array_queue.h
@@ -38,13 +38,13 @@ typedef struct aeron_spsc_concurrent_array_queue_stct
     }
     consumer;
 
-    uint64_t capacity;
-    uint64_t mask;
+    size_t capacity;
+    size_t mask;
     volatile void **buffer;
 }
 aeron_spsc_concurrent_array_queue_t;
 
-int aeron_spsc_concurrent_array_queue_init(aeron_spsc_concurrent_array_queue_t *queue, uint64_t length);
+int aeron_spsc_concurrent_array_queue_init(aeron_spsc_concurrent_array_queue_t *queue, size_t length);
 
 int aeron_spsc_concurrent_array_queue_close(aeron_spsc_concurrent_array_queue_t *queue);
 
@@ -73,7 +73,7 @@ inline aeron_queue_offer_result_t aeron_spsc_concurrent_array_queue_offer(
         queue->producer.head_cache = current_head;
     }
 
-    const uint64_t index = current_tail & queue->mask;
+    const size_t index = (size_t)(current_tail & queue->mask);
 
     AERON_PUT_ORDERED(queue->buffer[index], element);
     AERON_PUT_ORDERED(queue->producer.tail, current_tail + 1);
@@ -84,7 +84,7 @@ inline aeron_queue_offer_result_t aeron_spsc_concurrent_array_queue_offer(
 inline volatile void *aeron_spsc_concurrent_array_queue_poll(aeron_spsc_concurrent_array_queue_t *queue)
 {
     const uint64_t current_head = queue->consumer.head;
-    const uint64_t index = current_head & queue->mask;
+    const size_t index = (size_t)(current_head & queue->mask);
 
     volatile void *item;
     AERON_GET_VOLATILE(item, queue->buffer[index]);
@@ -98,8 +98,8 @@ inline volatile void *aeron_spsc_concurrent_array_queue_poll(aeron_spsc_concurre
     return item;
 }
 
-inline uint64_t aeron_spsc_concurrent_array_queue_drain(
-    aeron_spsc_concurrent_array_queue_t *queue, aeron_queue_drain_func_t func, void *clientd, uint64_t limit)
+inline size_t aeron_spsc_concurrent_array_queue_drain(
+    aeron_spsc_concurrent_array_queue_t *queue, aeron_queue_drain_func_t func, void *clientd, size_t limit)
 {
     uint64_t current_head = queue->consumer.head;
     uint64_t next_sequence = current_head;
@@ -107,7 +107,7 @@ inline uint64_t aeron_spsc_concurrent_array_queue_drain(
 
     while (next_sequence < limit_sequence)
     {
-        const uint64_t index = next_sequence & queue->mask;
+        const size_t index = (size_t)(next_sequence & queue->mask);
         volatile void *item;
         AERON_GET_VOLATILE(item, queue->buffer[index]);
 
@@ -125,7 +125,7 @@ inline uint64_t aeron_spsc_concurrent_array_queue_drain(
     return next_sequence - current_head;
 }
 
-inline uint64_t aeron_spsc_concurrent_array_queue_drain_all(
+inline size_t aeron_spsc_concurrent_array_queue_drain_all(
     aeron_spsc_concurrent_array_queue_t *queue, aeron_queue_drain_func_t func, void *clientd)
 {
     uint64_t current_head = queue->consumer.head;
@@ -135,7 +135,7 @@ inline uint64_t aeron_spsc_concurrent_array_queue_drain_all(
     return aeron_spsc_concurrent_array_queue_drain(queue, func, clientd, current_tail - current_head);
 }
 
-inline uint64_t aeron_spsc_concurrent_array_queue_size(aeron_spsc_concurrent_array_queue_t *queue)
+inline size_t aeron_spsc_concurrent_array_queue_size(aeron_spsc_concurrent_array_queue_t *queue)
 {
     uint64_t current_head_before;
     uint64_t current_tail;
@@ -151,7 +151,7 @@ inline uint64_t aeron_spsc_concurrent_array_queue_size(aeron_spsc_concurrent_arr
     }
     while (current_head_after != current_head_before);
 
-    uint64_t size = current_tail - current_head_after;
+    size_t size = (size_t)(current_tail - current_head_after);
     if ((int64_t)size < 0)
     {
         return 0;
